chore: move resolution error to errors
